+++ src/components/units/content.indicator/content.indicator.comp.tsx
