+++ src/components/units/content.indicator/content.indicator.style.ts
