+++ src/components/units/multicolor.text/multicolor.text.style.ts
