+++ src/components/units/trending.Item/trending.item.style.ts
