+++ src/pages/index.tsx
@@ -1,9 +1,9 @@
 import { Box, Divider, makeStyles, TextField, Typography } from "@material-ui/core";
 import LayoutLandingPage from '../layout/layout.landingPage/layout.landingPage';
 import heroImgURL from '../assets/landingPage/heroimg.png';
-import MulticolorText from "../components/multicolor.text/multicolor.text.comp";
+import MulticolorText from "../components/units/multicolor.text/multicolor.text.comp";
 import Trending from '../components/trending/trending.comp';
-import TrendingItem from '../components/trending.Item/trending.item.comp';
+import TrendingItem from '../components/units/trending.Item/trending.item.comp';
 import style from '../style/index.style';
 
 const imgUrl = 'https://images.unsplash.com/photo-1517935706615-2717063c2225?ixlib=rb-1.2.1&ixid=eyJhcHBfaWQiOjEyMDd9&auto=format&fit=crop&w=802&q=80';
